requirements.txt
streamlit
google-generativeai
PyPDF2
pandas
python-dotenv
extra-streamlit-components
